# swagger.yml
openapi: 3.0.0
info:
  title: Quinnn API List
  description: This is a reference api list of Quinnn Backend.
  version: 1.0.0
  contact:
    email: ankuragarwa246@gmail.com
servers:
  - url: https://lgbtq-backend-free.azurewebsites.net/api
    description: Production Server (Azure)
  - url: http://localhost:9999/api
    description: Development server
paths:
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: Signup
      requestBody:
        $ref: '#/components/requestBodies/Signup'
      responses:
        '400':
          description: Invalid input
        '500':
          description: Internal Server Error
        '200':
          description: Admin Created Successfully
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User Login
      requestBody:
        $ref: '#/components/requestBodies/Login'
      responses:
        '400':
          description: Invalid input
        '500':
          description: Internal Server Error
        '200':
          description: User Logged In Successfully
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Forgot Password
      requestBody:
        $ref: '#/components/requestBodies/ForgotPassword'
      responses:
        '200':
          description: Password Reset Link Sent Successfully
        '400':
          description: Invalid input
        '404':
          description: Admin Not Found
        '500':
          description: Internal Server Error
...
  /auth/reset-password/{token}:
    get:
      tags:
        - Authentication
      summary: Verify the link clicked by the user via email (From frontend)
      parameters:
        - in: path
          name: token
          description: Reset Password Token
      responses:
        '200':
          description: Verified, now you can enter new password
        '404':
          description: Admin Not Found
        '500':
          description: Internal Server Error
...
components:
  schemas:
...
    SignupInput:
      type: object
      properties:
        name:
          type: string
          required: true
        username:
          type: string
          uniqueItems: true
          required: true
        email:
          type: string
          uniqueItems: true
          required: true
        password:
          type: string
          required: true
        gender:
          type: string
          required: true
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    ForgotPasswordInput:
      type: object
      properties:
        email:
          type: string
...
  securitySchemes:
    bearerAuth:
      name: Authorization
      type: http
      scheme: bearer
      in: header
      bearerFormat: JWT
  requestBodies:
    Signup:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SignupInput'
        description: Signup
    Login:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LoginInput'
        description: Login
    ForgotPassword:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ForgotPasswordInput'
        description: Forgot Password
...